refactor: remove duplicated code for search
