+++ apps/storefront/src/components/composites/search-section.tsx
@@ -24,10 +24,10 @@ import {
    SelectValue,
 } from '@/components/ui/select'
 import { Slider } from '@/components/ui/slider'
+import { useSearchProducts } from '@/hooks/use-search-products'
 import { cn } from '@/lib/utils'
 import { Check, ChevronDown } from 'lucide-react'
-import { useRouter, useSearchParams } from 'next/navigation'
-import { startTransition, useRef, useState } from 'react'
+import { useState } from 'react'
 
 export type SearcSectionProps = {
    categories: string[]
@@ -35,82 +35,18 @@ export type SearcSectionProps = {
 }
 
 export function SearchSection({ categories, brands }: SearcSectionProps) {
-   const searchParams = useSearchParams()
-   const [priceRange, setPriceRange] = useState([
-      Number(searchParams.get('min_price') ?? '0'),
-      Number(searchParams.get('max_price') ?? '1000'),
-   ])
-   const [categoryOpen, setCategoryOpen] = useState(false)
    const [brandOpen, setBrandOpen] = useState(false)
-   const sliderDebounce = useRef<NodeJS.Timeout>(null)
-   const router = useRouter()
-
-   const selectedCategories = searchParams.getAll('category')
-   const selectedBrands = searchParams.getAll('brand')
-
-   function handleSearchWithMultiQueryParam(key: string, value: string) {
-      const newSearchParams = new URLSearchParams(searchParams.toString())
-      const currentMultiValues = newSearchParams.getAll(key)
-
-      if (currentMultiValues.includes(value)) {
-         newSearchParams.delete(key)
-         currentMultiValues
-            .filter((queryValue) => queryValue !== value)
-            .forEach((filteredValue) =>
-               newSearchParams.append(key, filteredValue)
-            )
-      } else {
-         newSearchParams.append(key, value)
-      }
-      startTransition(() => {
-         router.replace(`?${newSearchParams.toString()}`, {
-            scroll: false,
-         })
-      })
-   }
-
-   function handleSearchWithQueryParam(key: string, value: string) {
-      const newSearchParams = new URLSearchParams(searchParams.toString())
-      newSearchParams.set(key, value)
-      startTransition(() => {
-         router.replace(`?${newSearchParams.toString()}`, {
-            scroll: false,
-         })
-      })
-   }
-
-   function handleSearchForPriceRangeWithText(range: [number, number]) {
-      const newSearchParams = new URLSearchParams(searchParams.toString())
-      newSearchParams.set('min_price', String(range[0]))
-      newSearchParams.set('max_price', String(range[1]))
-
-      setPriceRange(range)
-
-      startTransition(() => {
-         router.replace(`?${newSearchParams.toString()}`, {
-            scroll: false,
-         })
-      })
-   }
-
-   function handleSearchForPriceRangeWithSlider(range: [number, number]) {
-      const newSearchParams = new URLSearchParams(searchParams.toString())
-      newSearchParams.set('min_price', String(range[0]))
-      newSearchParams.set('max_price', String(range[1]))
-      setPriceRange(range)
-
-      if (sliderDebounce.current) {
-         clearTimeout(sliderDebounce.current)
-      }
-
-      sliderDebounce.current = setTimeout(() => {
-         startTransition(() => {
-            router.replace(`?${newSearchParams.toString()}`, {
-               scroll: false,
-            })
-         })
-      }, 500)
-   }
+   const [categoryOpen, setCategoryOpen] = useState(false)
+   const {
+      handleSearchForPriceRangeWithSlider,
+      handleSearchForPriceRangeWithText,
+      handleSearchWithMultiQueryParam,
+      handleSearchWithQueryParam,
+      priceRange,
+      searchParams,
+      selectedBrands,
+      selectedCategories,
+   } = useSearchProducts()
 
    return (
       <div className="w-full mx-auto">
